API update: use box_region_aligned_alloc() directly
We decided to get rid of the box_region_alloc_array() helper in the
module API. It is better to comment box_region_alloc() and
box_region_aligned_alloc() properly and give hints when one or another
should be used.

What also doubts me about box_region_aligned_alloc(): it does not check
arguments for integer overflow, however one may expect that it'll do so,
because calloc() does.

See [1] for details.

[1]: https://lists.tarantool.org/pipermail/tarantool-patches/2020-October/019971.html