Fix hot reload
The scenario is the following:

 | require('tuple.keydef')
 | package.loaded['tuple.keydef'] = nil
 | require('tuple.keydef')

Before this commit we called ffi.metatype() for <struct tuple_keydef> on
each (re)load of the module. LuaJIT does not allow to call it twice for
the same type: it raises a Lua error and reloading of the module fails.

The fix is implemented in the way that does not break ability to declare
<struct tuple_keydef> using ffi.cdef() prior to a first load of the
module. One may need it to implement some general code that uses
ffi.istype() on a particular set of known types. Or for something else.
Anyway, it does not look as something definitely invalid, so I prefer to
stay on the safe side.

A couple of words about test cases. I added a test case that verifies
that the module is reloaded without errors, when there are alive links
to the old module table, as well as the opposite case: when all such
links are collected. In fact, the module don't handle any of such cases
specifically, so it is redundant. Just for the sake of completeness of
the testing.

Fixes #9