Don't spoil _G.tuple global variable
It was my mistake. As I remember, I just missed ability to pass the
module table into postload.lua without calling luaL_register() in the
form that sets package.loaded prior returning from the
luaopen_tuple_keydef() function. I don't know why, but I didn't think
that it'll set the global variable in addition to the package.loaded
field (or didn't pay enough attention to that).

I guess, setting a global variable is not what expected from an external
module, when it is loaded. Especially, when the term ('tuple') is so
common.