cmake: fail early if no module.h found
Commit f7f84f02a7c177c513dec8786f1bc59a200eafaa ('cmake: fix warning')
causes a little regression: if there is no tarantool/module.h header, we
should fail at the cmake stage. After the fix of the warning we'll fail
only at the building stage (when `make` is called). Here I return the
old behaviour (fail on the `cmake` stage) back.

Follows up #18