ci: add pure cmake . && make build flow
The new CI rule ensures that a usual developer build flow is working. It
also improves testing by using the `Debug` build type, which adds the
`-Werror` compiler option. Compiler warnings are interpreted as errors
in this case.

It is inspired by #24, where the `Debug` build fails, but
`RelWithDebInfo` is fine.

This commit follows a spirit of [1], where it was declared that we
should verify as developer's flow as well as user's flow. We missed
developer's flow in CI of this module and now it is here.

[1]: https://github.com/tarantool/tarantool/issues/6056